pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "avdp-synth"
version = "0.1.0"
description = "Synthetic dataset generation framework for the Audio Violence Dataset Project (AVDP)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # --- Config & validation ---
    "pydantic>=2.6",           # Config models, label schema validation
    "pyyaml>=6.0",             # YAML config loading
    "jinja2>=3.1",             # Scene script templates

    # --- Audio processing ---
    "soundfile>=0.12",         # WAV I/O (libsndfile bindings)
    "pydub>=0.25",             # Audio manipulation, mixing, segment slicing
    "torchaudio>=2.2",         # Resampling, spectral ops, Wiener filter
    "torch>=2.2",              # torchaudio dependency
    "pyroomacoustics>=0.7",    # Room impulse response simulation
    "scaper>=1.6",             # Structured soundscape synthesis (Tier C)
    "numpy>=1.26",
    "scipy>=1.12",             # Butterworth filter, spectral subtraction

    # --- TTS providers ---
    "azure-cognitiveservices-speech>=1.36",   # Azure TTS he-IL (primary)
    "google-cloud-texttospeech>=2.16",        # Google Cloud TTS he-IL (evaluation)

    # --- LLM script generation ---
    "anthropic>=0.25",         # Claude API for script generation
    "openai>=1.30",            # GPT-4 fallback for script generation

    # --- Metadata & I/O ---
    "pandas>=2.2",             # Manifest CSV generation and QA
    "jsonlines>=4.0",          # JSONL label file I/O

    # --- Hashing & caching ---
    "joblib>=1.3",             # TTS render cache (memory-mapped)
    "xxhash>=3.4",             # Fast content hashing for cache keys

    # --- CLI ---
    "click>=8.1",              # CLI entry points
    "rich>=13.7",              # Progress bars, console output

    # --- Dataset QA ---
    "librosa>=0.10",           # Audio analysis, SNR estimation, spectrogram QA
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",               # Linting + formatting
    "mypy>=1.9",
    "ipykernel>=6.29",         # Notebook QA
    "matplotlib>=3.8",         # QA visualisations
    "seaborn>=0.13",
]

[project.scripts]
avdp = "avdp.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["avdp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
